requirements.txt
# requirements.txt
pyrogram==2.0.30
pytgcalls==3.0.0
tgcrypto==1.2.3
python-dotenv==1.0.0
numpy==1.24.0
scipy==1.10.0
soundfile==0.12.0
pydub==0.25.1
librosa==0.10.0